+++ internal/model/usage.go
@@ -12,7 +12,7 @@ type Usage struct {
 	Hidden      bool           `json:"hidden"`
 	Score       int            `json:"score"`
 	Flag        string         `json:"flag"`
-	Attempt     int            `json:"attempt" gorm:"default:0"`
+	Attempt     int64          `json:"attempt" gorm:"default:0"`
 	Hints       string         `json:"hints"`
 	Tags        string         `json:"tags"`
 	CreatedAt   time.Time      `json:"-"`
